tts_server/requirements.txt
flask
elevenlabs==0.2.24
python-dotenv